Moved out batching client from storage client. Implemented multiple batch uploads aka the bunch upload.
