+++ Sync/StorageClient.swift
@@ -91,18 +91,6 @@ public class RecordTooLargeError: MaybeErrorType {
     }
 }
 
-public class BatchingNotSupported: MaybeErrorType {
-    public let response: POSTResult
-
-    public var description: String {
-        return "Sync server does not support batching."
-    }
-
-    public init(response: POSTResult) {
-        self.response = response
-    }
-}
-
 /**
  * Raised when the storage client is refusing to make a request due to a known
  * server backoff.
@@ -626,193 +614,6 @@ public class Sync15StorageClient {
     func clientForCollection<T: CleartextPayloadJSON>(collection: String, encrypter: RecordEncrypter<T>) -> Sync15CollectionClient<T> {
         let storage = self.serverURI.URLByAppendingPathComponent("storage", isDirectory: true)
         return Sync15CollectionClient(client: self, serverURI: storage, collection: collection, encrypter: encrypter)
-    }
-}
-
-
-typealias BatchUploadFunction = (lines: [String], ifUnmodifiedSince: Timestamp?, queryParams: [NSURLQueryItem]?) -> Deferred<Maybe<StorageResponse<POSTResult>>>
-
-public class Sync15BatchClient<T: CleartextPayloadJSON> {
-    private let config: InfoConfiguration
-
-    private var records: [Record<T>] = []
-    private let ifUnmodifiedSince: Timestamp?
-    private let uploader: BatchUploadFunction
-    private let serializeRecord: (Record<T>) -> String?
-
-    // TODO: use I-U-S.
-    // Each time we do the storage operation, we might receive a backoff notification.
-    // For a success response, this will be on the subsequent request, which means we don't
-    // have to worry about handling successes and failures mixed with backoffs here.
-
-    init(config: InfoConfiguration, ifUnmodifiedSince: Timestamp? = nil, serializeRecord: (Record<T>) -> String?, uploader: BatchUploadFunction) {
-        self.config = config
-        self.ifUnmodifiedSince = ifUnmodifiedSince
-        self.uploader = uploader
-        self.serializeRecord = serializeRecord
-    }
-
-    public func addRecords(records: [Record<T>]) {
-        log.debug("Adding \(records.count) records into batch")
-        self.records += records
-    }
-
-    public func commit(onCollectionUploaded: (POSTResult -> Void)) -> Success {
-
-        if records.isEmpty {
-            return succeed()
-        }
-
-        // Need to deduce how big the data is we're sending across using this reducer
-        let sizeReducer: (total: ByteCount, record: Record<T>) -> ByteCount = { total, record in
-            return total + (self.serializeRecord(record) ?? "").utf8.count
-        }
-
-        let sizeOfRecords = records.reduce(0, combine: sizeReducer)
-
-        // We have too many records for a single post so we'll either need to upload in a single batch or
-        // multiple batches
-        if records.count > config.maxTotalRecords || sizeOfRecords > config.maxTotalBytes {
-            // TODO: Break into multiple batches
-            return succeed()
-        }
-
-        if records.count > config.maxPostRecords || sizeOfRecords > config.maxPostBytes {
-            return self.batchUpload(records, ifUnmodifiedSince: self.ifUnmodifiedSince, onCollectionUploaded: onCollectionUploaded)
-        }
-
-        // We can just do a single post instead of batching
-        log.debug("Batch fits within a single request. Submitting records in a single post.")
-        let lines = optFilter(records.map(self.serializeRecord))
-        return self.uploader(lines: lines, ifUnmodifiedSince: self.ifUnmodifiedSince, queryParams: nil)
-            >>== effect({ onCollectionUploaded($0.value) })
-            >>> succeed
-    }
-
-    public func batchUpload(records: [Record<T>], ifUnmodifiedSince: Timestamp?, onCollectionUploaded: (POSTResult -> Void)) -> Success {
-        let batchingResult = batchesFromRecords(records)
-        guard var batches = batchingResult.successValue else {
-            let failure = batchingResult.failureValue!
-            log.debug("Unable to generate batches from records submitted to batch client: \(failure)")
-            return deferMaybe(failure)
-        }
-
-        let firstBatch = batches.removeFirst()
-        let lastBatch = batches.last ?? []
-
-        return startBatch(firstBatch).bind { result in
-            guard let token = result.successValue else {
-                // Check if we didn't get a token back/batching isn't supported and push up records using regular single posts
-                if let error = result.failureValue as? BatchingNotSupported {
-
-                    // Since we uploaded the first set already as part of the start call and we don't support
-                    // batching, make sure to invoke the upload callback to let the others know we committed
-                    // new records to the collection server-side
-                    onCollectionUploaded(error.response)
-
-                    // Walk through the batches, posting along the way and invoking onUpload
-                    let perChunk: (lines: [String]) -> Success = { lines in
-                        return self.uploader(lines: lines, ifUnmodifiedSince: ifUnmodifiedSince, queryParams: nil)
-                            >>== effect({ onCollectionUploaded($0.value) })
-                            >>> succeed
-                    }
-                    return walk(batches, f: perChunk)
-                } else {
-                    // Bubble up other errors
-                    return deferMaybe(result.failureValue!)
-                }
-            }
-
-            // Remove the last batch - we handle the last call in a special case.
-            batches.removeLast()
-
-            // When batching, each upload in the batch is uploaded to a temporary collection until we specif
-            // commit=true. At this point, the temporary collection is pushed to the real collection on the server.
-            // It is at this point we want to say the collection has been uploaded.
-            return self.uploadBatches(token, batches: batches) >>> {
-                return self.finishBatch(token, lines: lastBatch)
-                    >>== effect({ onCollectionUploaded($0.value) })
-                    >>> succeed
-            }
-        }
-    }
-
-    private func startBatch(lines: [String]) -> Deferred<Maybe<BatchToken>> {
-        let batchStartParam = NSURLQueryItem(name: "batch", value: "true")
-
-        // Attempt to upload some records and see if we get back a token we can use for batches.
-        return self.uploader(lines: lines, ifUnmodifiedSince: ifUnmodifiedSince, queryParams: [batchStartParam]) >>== { storageResponse in
-            if let batchToken = storageResponse.value.batchToken {
-                log.debug("Uploaded \(lines.count) records and received batch token \(batchToken)")
-                return deferMaybe(batchToken)
-            } else {
-                log.debug("Uploaded \(lines.count) records but received no batch token")
-                return deferMaybe(BatchingNotSupported(response: storageResponse.value))
-            }
-        }
-    }
-
-    private func uploadBatches(token: BatchToken, batches: [[String]]) -> Success {
-        let batchParam = NSURLQueryItem(name: "batch", value: token)
-        let uploadBatch: (lines: [String]) -> Success = { lines in
-            return self.uploader(lines: lines, ifUnmodifiedSince: self.ifUnmodifiedSince, queryParams: [batchParam])
-                >>> effect({ log.debug(("Uploaded \(lines.count) records for batch \(token)")) })
-        }
-
-        return walk(batches, f: uploadBatch)
-    }
-
-    private func finishBatch(token: BatchToken, lines: [String]) -> Deferred<Maybe<StorageResponse<POSTResult>>> {
-        let batchParam = NSURLQueryItem(name: "batch", value: token)
-        let commitParam = NSURLQueryItem(name: "commit", value: "true")
-        return self.uploader(lines: lines, ifUnmodifiedSince: ifUnmodifiedSince, queryParams: [batchParam, commitParam])
-    }
-
-    func batchesFromRecords(records: [Record<T>]) -> Maybe<[[String]]> {
-
-        // Place the record ID alongside the line text for error reporting later on
-        let idWithLine: (record: Record<T>) -> (GUID, String)? = { record in
-            guard let line = self.serializeRecord(record) else {
-                return nil
-            }
-            return (record.id, line)
-        }
-
-        let pairs : [(GUID, String)] = records.flatMap(idWithLine)
-
-        var batches = [[String]]()
-        var batch = [String]()
-        var size = 0
-
-        for (id, line) in pairs {
-            let lineSize = line.utf8.count
-
-            if lineSize > Sync15StorageClient.maxRecordSizeBytes {
-                return Maybe(failure: RecordTooLargeError(size: lineSize, guid: id))
-            }
-
-            // If adding the line keeps the batch size and number of records below the limits, proceed.
-            // Otherwise, add the filled batch to our batches and start a new one.
-
-            // Add 1 to account for the newline
-            guard (size + lineSize + 1) <= config.maxPostBytes && (batch.count + 1) <= config.maxPostRecords else {
-                // Start the next batch
-                batches.append(batch)
-                size = lineSize
-                batch = [line]
-                continue
-            }
-
-            size += (lineSize + 1)
-            batch.append(line)
-        }
-
-        // Don't forget to add the last batch!
-        if !batch.isEmpty {
-            batches.append(batch)
-        }
-
-        return Maybe(success: batches)
     }
 }
